update script to download and load multuple days, cleaning with python, filtering with polars

+++ clean_raw_input.py
@@ -1,25 +1,25 @@
 import json
 import glob
+from pathlib import Path
 
 
 def clean_ndjson_file(input_file, output_file):
+    if Path(output_file).exists():
+        print("... skipping!!!")
+        return
+
     with open(input_file, "r") as infile, open(output_file, "w") as outfile:
         for line in infile:
             try:
                 # Attempt to parse each line as a JSON object
                 data = json.loads(line)
 
-                # Example cleanup: If the object is invalid, replace with None or clean accordingly
-                if not isinstance(
-                    data, (dict, list, str, int, float, bool, type(None))
-                ):
-                    data = None  # Handle as per your requirement
-
                 # Write the cleaned line back to the output file
                 outfile.write(json.dumps(data) + "\n")
             except json.JSONDecodeError as e:
                 print(f"Invalid JSON in line: {e}")
                 # Optionally, you can skip the invalid lines or log them
+        print("... done!!!")
 
 
 files = glob.glob("raw_input/*.json")
@@ -29,6 +29,6 @@ if __name__ == "__main__":
 
         filename = input_file.split("/")[-1]
         output_file = f"clean_input/{filename}"
-        print(f"processing file {i}: {input_file} -> {output_file}")
+        print(f"processing file {i}: {input_file} -> {output_file}", end="")
         clean_ndjson_file(input_file, output_file)
     print("Done!!!")
